Added CEX to the list of sites to browse.

+++ Objects/Product.cs
@@ -2,7 +2,7 @@
 {
     public class Product
     {
-        public int Price;
+        public double Price;
         public string Url;
         public string RetailName;
     }
